requirements.txt
fastapi
uvicorn
numpy
pillow
tensorflow
python-multipart
requests
